requirements.txt
aiohttp==3.9.2
aiosignal==1.3.1
alembic==1.11.3
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.15.0
blinker==1.6.2
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
chess==1.10.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
courlan==1.2.0
cycler==0.12.1
dateparser==1.2.0
distro==1.9.0
emoji==1.7.0
exceptiongroup==1.2.1
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
fonttools==4.43.1
frozenlist==1.4.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.111.0
greenlet==2.0.2
gspread==5.12.0
gunicorn==21.2.0
h11==0.14.0
htmldate==1.8.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jusText==3.0.1
kiwisolver==1.4.5
lxml==5.2.2
lxml_html_clean==0.1.1
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib==3.8.1
multidict==6.0.4
numpy==1.26.1
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.35.3
packaging==23.2
Pillow==10.1.0
pilmoji==2.0.4
psycopg2-binary==2.9.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.7.4
pydantic_core==2.18.4
pydub==0.25.1
pygame==2.5.2
PyJWT==2.8.0
PyMuPDF==1.23.21
PyMuPDFb==1.23.9
pyparsing==3.1.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
regex==2024.5.15
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.20
stripe==8.2.0
tld==0.13
tqdm==4.66.1
trafilatura==1.10.0
twilio==8.10.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.4
waitress==3.0.0
Werkzeug==2.3.7
yarl==1.9.2
boto3==1.34.34
assemblyai==0.19.1
faiss-cpu==1.7.4
bs4==0.0.1
elevenlabs==0.2.26